Revert "Revert "created routing with where context gets value based on url param id""
