+++ src/pages/ChatPage/ChatPage.tsx
@@ -1,48 +1,15 @@
-import { useEffect } from 'react'
-import { useUserContext } from '../../context/auth'
-import { useChatContext } from '../../context/chat'
-import { ChatDocument, convertChat, UserDocument } from '../../types'
-import { doc, getDoc, onSnapshot } from 'firebase/firestore'
-import { db } from '../../firebase'
+import { Route, Routes } from 'react-router-dom'
 import Sidebar from '../../components/Sidebar/Sidebar'
 import ChatWindow from '../../components/ChatWindow/ChatWindow'
 import './ChatPage.css'
 
 const ChatPage = () => {
-  const [currentUser] = useUserContext()
-  const [, dispatch] = useChatContext()
-
-  useEffect(() => {
-    const getChatsOnUserChange = (currentUserId: string) => {
-      const unsub = onSnapshot(
-        doc(db, 'users', currentUserId),
-        async (userSnap) => {
-          const userDoc = userSnap.data() as UserDocument
-          const chatRefs = userDoc.chatRefs
-          const chatSnaps = await Promise.all(
-            chatRefs.map((chatRef) => getDoc(chatRef))
-          )
-
-          const chats = chatSnaps
-            .filter((chatSnap) => chatSnap.exists())
-            .map((chatSnap) => {
-              const chatDoc = chatSnap.data() as ChatDocument
-              return convertChat(currentUserId, chatDoc)
-            })
-
-          dispatch({ type: 'GET_CHATS', payload: chats })
-        }
-      )
-      return () => unsub()
-    }
-
-    getChatsOnUserChange(currentUser.uid)
-  }, [currentUser, dispatch])
-
   return (
     <div className="chat-page">
       <Sidebar />
-      <ChatWindow />
+      <Routes>
+        <Route path="/chat/:id" element={<ChatWindow />} />
+      </Routes>
     </div>
   )
 }
